EMAR-231:  Rewrote the PatientRepository calls for Image, and  - Removed the EF-code foreign key between sites and external_ids  - Reverted all the changes of the site_id int --> long  - Added ability to retrieve the site_id by site_name (to facilitate PostMan tests)
git-tfs-id: [http://ros-tfs-px01.picis.com:8080/tfs/emar]$/;C363

+++ WebAPI/API/C_Development/C_DevTrunk/Emar.Core/Patients/Model/PatientDto.cs
@@ -92,7 +92,7 @@ namespace Emar.Core.Patients.Model
         public decimal? WeightInKg { get; set; }
 
         #region geography - room, ward, department
-        public long SiteId { get; set; }
+        public int SiteId { get; set; }
 
         string departmentCode;
         public string DepartmentCode
